add drop function support and create a new test case file
